# docs/swagger.yaml
definitions:
...
  domain.LayoutInfo:
    properties:
      layout_id:
        type: string
      params:
        items:
          $ref: '#/definitions/domain.Param'
        type: array
    type: object
...
  domain.Param:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
  domain.ParamScreenUpd:
    properties:
      device_id:
        type: string
      screenshot_status:
        type: string
      screenshot_time:
        type: string
    type: object
  domain.ParamsScreenUpd:
    items:
      $ref: '#/definitions/domain.ParamScreenUpd'
    type: array
...
  domain.Screenshot:
    properties:
      created_at:
        description: timestamptz not null default current_timestamp, -- дата-время
          создания записи
        type: string
      creator:
        description: varchar(128) not null default current_user, -- пользователь,
          создавший запись
        type: string
      device_id:
        description: text not null, -- идентификатор устройства, с которого получен
          скриншот
        type: string
      layout_id:
        description: text null, -- идентификатор схемы размещения, которой принадлежит
          устройство с которого получен скриншот
        type: string
      layout_info:
        $ref: '#/definitions/domain.LayoutInfo'
        description: jsonb not null default '{}'::jsonb, -- более детальная проектная
          информация устройства
        type: object
      notes:
        description: text null, -- заметки
        type: string
      screenshot_status:
        description: 'статус скриншота: new-по умолчанию, processed-блокировка от
          изменений другими пользователями (не более 1 минуты), to_delete-помечен
          для удаления, archived-запись для вечного хранения'
        type: string
      screenshot_time:
        description: stamptz not null, -- дата-время создания скриншота
        type: string
      store_id:
        description: text not null, -- идентификатор магазина, в котором находилось
          устройство с которого получен скриншот
        type: string
      url:
        description: text not null, -- адрес местонахождения скриншота
        type: string
      url_aliases:
        description: api layer only, alt links to screenshots
        items:
          type: string
        type: array
    type: object
  domain.Screenshots:
    items:
      $ref: '#/definitions/domain.Screenshot'
    type: array
...
  domain.ZoneDataEvaluation:
    properties:
      comment:
        type: string
      is_full:
        type: boolean
      layout_id:
        type: string
      record_time:
        type: string
      service_channel_block_id:
        type: string
      store_id:
        type: string
    type: object
  domain.ZoneDataEvaluations:
    items:
      $ref: '#/definitions/domain.ZoneDataEvaluation'
    type: array
...
  infra.ErrResponse:
    properties:
      code:
        description: application-specific error code
        type: integer
      error:
        description: application-level error message, for debugging
        type: string
      status:
        description: user-level status message
        type: string
    type: object
  infra.HTTPError:
    properties:
      message:
        type: object
    type: object
...
  infra.ResultSet:
    properties:
      count:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  infra.ScreenshotsResponse:
    properties:
      data:
        $ref: '#/definitions/domain.Screenshots'
        type: object
      result_set:
        $ref: '#/definitions/infra.ResultSet'
        type: object
    type: object
  infra.SuccessResponse:
    properties:
      httpcode:
        description: http response status code
        type: integer
      message:
        description: text of message
        type: string
      status:
        description: user-level status message
        type: string
    type: object
...
info:
  contact:
    email: 1020@watcom.ru
    name: API Support
    url: https://helpdesk.watcom.ru
  description: |-
    This is a general service for interacting with layout configuration.
    Общий сервис для взаимодействия с конфигурацией схем размещения.
  license: {}
  title: LayoutConfig API
  version: 2.0/1.1.21
paths:
...
  /v2/data/queue/evaluations:
    get:
      description: |-
        get zone data evaluation for serviceChannel blocks by intervals
        layout_id (recommended parameter), if not pass datasource may be not correct
        can be filtered by layout_id, store_id, parent_zone_id and daterange parameters from and to.
        service_channel_block_id is serviceChannel block identifier
        from/to can be: YYYY-MM-DDTHH:mm:ss+07:00 or naive YYYY-MM-DD HH:mm:ss then the server's local timezone is applied
      parameters:
      - description: uuid format, default=*
        in: path
        name: layout_id
        required: true
        type: string
      - description: default=*
        in: query
        name: store_id
        type: string
      - description: default=*
        in: query
        name: service_channel_block_id
        type: string
      - description: default=*
        in: query
        name: is_full
        type: string
      - description: ISO8601 datetime, default begin of day
        in: query
        name: from
        type: string
      - description: ISO8601 datetime, dafault current time
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ZoneDataEvaluations'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/infra.ErrResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/infra.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/infra.ErrResponse'
      summary: Get zone data evaluation for serviceChannel blocks
      tags:
      - data/queue
...
  /v2/screenshots:
    get:
      description: |-
        get slice of screenshots with layout_id, store_id, zone_id, device_id, status, offset, limit parameters
        fields - comma separated values of field names, can be layout_id,store_id,device_id,url,notes... all of them described in the model
        screenshot_status can be new - default processed - blocked for changes, to_delete - marked for delete, archived - marked for all time store
        from/to can be: YYYY-MM-DDTHH:mm:ss+07:00 or naive YYYY-MM-DD HH:mm:ss then the server's local timezone is applied
      parameters:
      - description: default=*
        in: query
        name: layout_id
        type: string
      - description: default=*
        in: query
        name: store_id
        type: string
      - description: default=*
        in: query
        name: zone_id
        type: string
      - description: default=*
        in: query
        name: device_id
        type: string
      - description: default=*
        in: query
        name: screenshot_status
        type: string
      - description: default=0
        in: query
        name: offset
        type: integer
      - description: default=20
        in: query
        name: limit
        type: integer
      - description: ISO8601 datetime, default begin of day
        in: query
        name: from
        type: string
      - description: ISO8601 datetime, dafault current time
        in: query
        name: to
        type: string
      - description: layout_id,store_id,device_id..default=all
        in: query
        name: fields
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/infra.ScreenshotsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/infra.ErrResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/infra.ErrResponse'
      summary: Get screenshots from the storage
      tags:
      - screenshots
    put:
      consumes:
      - application/json
      description: |-
        Updates screenshots status
        example upd screenshots statuses:
        [{
        "device_id": "1234567",
        "screenshot_time": "2020-09-01T00:15:15+03:00",
        "screenshot_status": "to_delete"
        },
        {
        "device_id": "1234568",
        "screenshot_time": "2020-09-01T00:15:15+03:00",
        "screenshot_status": "to_delete"
        }]
      parameters:
      - description: screenshots update status properties
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/domain.ParamsScreenUpd'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/infra.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/infra.ErrResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/infra.ErrResponse'
      summary: Updates screenshots status
      tags:
      - screenshots
  /v2/screenshots/attime:
    get:
      description: |-
        get slice of screenshots with layout_id, store_id, zone_id, device_id, at time momen parameters
        fields - comma separated values of field names, can be layout_id,store_id,device_id,url,notes... all of them described in the model
        at can be: YYYY-MM-DDTHH:mm:ss+07:00 or naive YYYY-MM-DD HH:mm:ss then the server's local timezone is applied
      parameters:
      - description: default=*
        in: query
        name: layout_id
        type: string
      - description: default=*
        in: query
        name: store_id
        type: string
      - description: default=*
        in: query
        name: zone_id
        type: string
      - description: default=*
        in: query
        name: device_id
        type: string
      - description: ISO8601 datetime, default=currentMoment-1min accuracy 15s
        in: query
        name: at
        type: string
      - description: layout_id,store_id,device_id..default=all
        in: query
        name: fields
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Screenshots'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/infra.ErrResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/infra.ErrResponse'
      summary: Get all screenshots from the storage on the specified time
      tags:
      - screenshots
...
swagger: "2.0"